+++ CHANGELOG.md
@@ -5,16 +5,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.0] - 2021-xx-xx
+This release contains mainly 2 topics:
+* **NGSI-LD Support (experimental)** - Starting with this release the experimental Orion LD-server (Version: post-v0.7) can be used with OpenSim.
+* **Changes in Options** - This release has some changes in options in a way, that there is now a more generic way for creating necessary headers. If you are familiar with the current options (in 1.0.0), here is, how to "migrate" to the new options:
+  * **Bearer token** - In 1.0.0 you might have used the -b [--bearer] option like: `... -b BEARER_TOKEN ...`. With this new release you will have to use `-H Authorization "Bearer INSERT_BEARER_TOKEN_HERE"` instead.
+  * **Tenant** - If you used the -t [--tenant] option like: `... -t MY_TENANT ...` to set the Fiware-service you have to switch to `-H Fiware-service MY_TENANT` now.
+  * **X-Gravitee-Api-Key** - If you used the -x [--x-api-key] option to set an X-Gravitee-Api-Key, please use `-H X-Gravitee-Api-Key MY_API_KEY` now.
+
+### Added
+- NGSI-LD support.
+- Option -H [--header] can be used to add header.
+
 ### Deleted
-- Removed options -b [--bearer] and -t [--tenant].
+- Removed options -b [--bearer], -x [--x-api-key] and -t [--tenant].
 
 ### Changed
-- Made -x [--x-api-key] more generic (-x [--header])
+- Nothing
 
 ## [1.0.0] - 2021-01-17
 ### Added
 - Version- and Copyright-printout.
 - CONTRIBUTION.md
+
+### Deleted
+- Nothing.
 
 ### Changed
 - Nothing.